refactor: reorganize metrics placement
