+++ libs/common/cache/cache.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { MetricsModule } from '../metrics/metrics.module';
+import { MetricsModule } from './infrastructure/metrics/metrics.module';
 import { RedisCacheService } from './infrastructure/cache.service';
 import { CacheService } from './application/cache.interface';
-import { MetricsService } from '../metrics/metrics.service';
+import { MetricsService } from './infrastructure/metrics/metrics.service';
 import { MetricsCacheDecorator } from './infrastructure/decorators/metrics-cache.decorator';
 import { ConfigModule } from '@nestjs/config';
