+++ libs/common/cache/infrastructure/metrics/metrics.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { makeCounterProvider } from '@willsoto/nestjs-prometheus';
 import { MetricsService } from './metrics.service';
-import { CACHE_METRICS_NAMES } from '../../../libs/constants/enums/metrics';
+import { CACHE_METRICS_NAMES } from '../../../../constants/enums/metrics';
 
 @Module({
   providers: [
